refactor multiPlayerTyping component by using custom hooks
