+++ app/room/[roomId]/_components/multiPlayerTyping.tsx
@@ -3,18 +3,17 @@ import Word from '@/components/word';
 import {
   calculateCompletionPercentage,
   cn,
-  generateWords,
-  parseConvexError,
   validCharacters,
 } from '@/lib/utils';
 import { useTypingStore } from '@/stores/typingStore';
-import { useRef, useEffect, useState, useCallback, use } from 'react';
+import { useRef, useEffect, useState, useCallback } from 'react';
 import { useConvex, useQuery } from 'convex/react';
 import { api } from '@/convex/_generated/api';
 import { useUserStore } from '@/stores/userStore';
-import { toast } from 'sonner';
 import CountdownSeparator from '@/app/room/[roomId]/_components/countDownSeperator';
-import confetti from 'canvas-confetti';
+import useTimer from '@/hooks/useTimer';
+import useWinner from '@/hooks/useWinner';
+import useResetNotification from '@/hooks/useResetNotification';
 
 interface MultiTypingTestProps {
   roomId: string;
@@ -26,7 +25,6 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
     typedWord,
     moveToNextWord,
     wordList,
-    setWordListLength,
     setStrictMode,
     setWordList,
     setTypedWord,
@@ -35,132 +33,45 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
   } = useTypingStore();
 
   const convexWordList = useQuery(api.room.getWordList, { roomId });
-
-  const [timeLeft, setTimeLeft] = useState<number>(0);
-  const [initialCountDown, setInitialCountDown] = useState(5);
-  const [isInitialCountDownRunning, setIsInitialCountDownRunning] =
-    useState(false);
   const convex = useConvex();
   const user = useUserStore((state) => state.user);
   const ownerId = useQuery(api.room.getOwner, { roomId });
   const isOwner = ownerId === user?.userId;
   const roomTimer = useQuery(api.room.getTimer, { roomId });
   const winner = useQuery(api.room.getWinner, { roomId });
-  const resetNotification = useQuery(api.room.getResetNotification, { roomId });
 
-  const timerStartedRef = useRef(false);
+  const {
+    timeLeft,
+    initialCountDown,
+    isInitialCountDownRunning,
+    setInitialCountDown,
+    setIsInitialCountDownRunning,
+    setTimeLeft,
+    startInitialCountDown,
+    startMainTimer,
+    resetRace,
+  } = useTimer({
+    roomId,
+    user,
+    isOwner,
+  });
+
+  const { sendResetNotification } = useResetNotification({
+    roomId,
+    resetRace,
+  });
 
   const inputRef = useRef<HTMLDivElement>(null);
   const restartButtonRef = useRef<HTMLButtonElement>(null);
   const activeWordRef = useRef<HTMLDivElement>(null);
   const [isFocused, setIsFocused] = useState(false);
 
-  const sendResetNotification = useCallback(async () => {
-    try {
-      await convex.mutation(api.room.sendResetNotification, { roomId });
-    } catch (error) {
-      toast.error(
-        `Failed to send reset notification: ${parseConvexError(error)}`,
-      );
-    }
-  }, [convex, roomId]);
+  const { calculateWinner, displayWinner } = useWinner({ roomId });
 
-  const resetRace = useCallback(async () => {
-    try {
-      await convex.mutation(api.room.resetTimer, {
-        roomId,
-      });
-      await convex.mutation(api.room.resetRoom, {
-        roomId,
-      });
-      setTimeLeft(0);
-      setInitialCountDown(5);
-      setIsInitialCountDownRunning(false);
-      resetTypingState(false);
-      timerStartedRef.current = false;
-    } catch (error) {
-      toast.error(`Failed to reset timer: ${parseConvexError(error)}`);
-    }
-  }, [roomId, convex, resetTypingState]);
-
-  const showConfetti = useCallback(() => {
-    const duration = 5 * 1000;
-    const animationEnd = Date.now() + duration;
-    const defaults = { startVelocity: 30, spread: 360, ticks: 60, zIndex: 0 };
-
-    const randomInRange = (min: number, max: number) =>
-      Math.random() * (max - min) + min;
-
-    const interval = window.setInterval(() => {
-      const timeLeft = animationEnd - Date.now();
-
-      if (timeLeft <= 0) {
-        return clearInterval(interval);
-      }
-
-      const particleCount = 50 * (timeLeft / duration);
-      confetti({
-        ...defaults,
-        particleCount,
-        origin: { x: randomInRange(0.1, 0.3), y: Math.random() - 0.2 },
-      });
-      confetti({
-        ...defaults,
-        particleCount,
-        origin: { x: randomInRange(0.7, 0.9), y: Math.random() - 0.2 },
-      });
-    }, 250);
+  const focusInput = useCallback(() => {
+    setIsFocused(true);
+    inputRef.current?.focus();
   }, []);
-
-  const displayWinner = useCallback(
-    (winnerName: string, isCurrentUser: boolean) => {
-      if (isCurrentUser) {
-        toast.success(`You won! Congratulations, ${winnerName}!`);
-        showConfetti();
-      } else {
-        toast.info(`${winnerName} won the game!`);
-      }
-    },
-    [showConfetti],
-  );
-
-  const calculateWinner = useCallback(async () => {
-    try {
-      await convex.mutation(api.room.calcWinnerWhenTimerEnds, { roomId });
-    } catch (error) {
-      toast.error(`Failed to calculate winner: ${parseConvexError(error)}`);
-    }
-  }, [convex, roomId]);
-
-  const startInitialCountDown = useCallback(async () => {
-    if (!user) return;
-    try {
-      await convex.mutation(api.room.startInitialCountDown, {
-        roomId,
-        userId: user.userId,
-      });
-      setWordListLength(30);
-      await convex.mutation(api.room.setWordList, {
-        roomId,
-        wordList: generateWords(30),
-      });
-    } catch (error) {
-      toast.error(`Failed to start countdown: ${parseConvexError(error)}`);
-    }
-  }, [convex, roomId, user, setWordListLength]);
-
-  const startMainTimer = useCallback(async () => {
-    if (!user || !isOwner || timerStartedRef.current) return;
-    timerStartedRef.current = true;
-    try {
-      await convex.mutation(api.room.startTimer, {
-        roomId,
-        userId: user.userId,
-      });
-    } catch (error) {
-      toast.error(`Failed to start timer: ${parseConvexError(error)}`);
-    }
-  }, [roomId, user, isOwner, convex]);
 
   useEffect(() => {
     if (initialCountDown === 0 && isInitialCountDownRunning) {
@@ -200,7 +111,12 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
       setInitialCountDown(5);
       setIsInitialCountDownRunning(false);
     }
-  }, [roomTimer, startMainTimer]);
+  }, [
+    roomTimer,
+    startMainTimer,
+    setInitialCountDown,
+    setIsInitialCountDownRunning,
+  ]);
 
   useEffect(() => {
     if (roomTimer?.endTime) {
@@ -224,7 +140,7 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
     } else {
       setTimeLeft(0);
     }
-  }, [roomTimer, calculateWinner]);
+  }, [roomTimer, calculateWinner, focusInput, setTimeLeft]);
 
   useEffect(() => {
     if (!winner) return;
@@ -234,13 +150,6 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
 
     resetRace();
   }, [winner, user, resetRace, isOwner, displayWinner]);
-
-  useEffect(() => {
-    if (resetNotification) {
-      toast.info('The room has been reset');
-      resetRace();
-    }
-  }, [resetNotification, resetRace]);
 
   useEffect(() => {
     if (activeWordRef.current) {
@@ -299,11 +208,6 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
     ],
   );
 
-  const focusInput = useCallback(() => {
-    setIsFocused(true);
-    inputRef.current?.focus();
-  }, []);
-
   return (
     <>
       <div className="flex flex-col items-center w-3/4 gap-3">
@@ -311,7 +215,6 @@ export default function MultiTypingTest({ roomId }: MultiTypingTestProps) {
           isInitialCountDownRunning={isInitialCountDownRunning}
           initialCountDown={initialCountDown}
           timeLeft={timeLeft}
-          timerStartedRef={timerStartedRef}
           startInitialCountDown={startInitialCountDown}
           resetRace={sendResetNotification}
           isOwner={isOwner}
